+++ src/onegov.town/HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 Unreleased
 ~~~~~~~~~~
 
+- Adds the ability to create reservations and to accept/reject them.
+  [href]
+
 - The edit links for the model shown on the ticket view are now only visible
   if the ticket is in 'pending' state. To change something on the model, the
   ticket needs to be accepted/reopened.